+++ addon/adapters/chronos-journal.js
@@ -1,0 +1,8 @@
+import OsfAdapter from './osf-adapter';
+
+export default OsfAdapter.extend({
+    namespace: '_/chronos',
+    pathForType: function () {
+        return 'journals';
+    }
+});