+++ app/components/lazy-options-load-more/component.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-osf/components/lazy-options-load-more/component';