+++ app/models/chronos-submission.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-osf/models/chronos-submission';