+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.24.0] - 2019-04-09
+### Added
+- `chronos-journal` model/adapter/serializer
+- `chronos-submission` model/adapter/serializer
+- `preprint-request` model/adapter/serializer
+- `preprint-request-action` model/adapter/serializer
+- `lazy-options-load-more` component
+
+# Changed
+- `preprint` model - added `requests` relationship
+- `lazy-options` component - moved load more button to its own component
+
+# Removed
+- `jquery-extensions` instance intializer
+- `jquery-checkinview` util
+
 ## [0.23.0] - 2019-03-04
 ### Added
 - `X-CSRFToken` header to AJAX requests